Workflow for Stale PR and Issues
This workflow will run every day at 00.00 UTC to check for any issues/ PRs that have been inactive for over 30 and will close them in another 15 days.

Additionally, the stale issues will be marked with `stale-issue` label and the stale PRs will be marked with `stale-pr` label. Issues/ PR having `awaiting-approval` or `work-in-progress` labels are excluded from this check.